Prepare shared flows for the debugger agent to support async stack traces
The agent needs three entities to establish a proper asynchronous stack traces connection:
  - a capture point -- method that indicates the stack trace that precedes the current stack trace;
  - an insertion point -- method within the current stack trace;
  - a key -- an object that is present in both points and is unique enough to bridge two points properly.

This change tweaks the code a bit to introduce the three entities in MutableSharedFlow and MutableStateFlow.

The key for MutableSharedFlow is the element itself. For MutableSharedFlow, the element is wrapped into a unique object to prevent bridging mistakes when two equal elements are emitted from different places.